+++ types/affinda.d.ts
@@ -751,7 +751,12 @@ export declare interface Annotation {
     rectangle: Rectangle | null;
     pageIndex: number | null;
     raw: string | null;
+    /** The overall confidence that the model's prediction is correct */
     confidence: number | null;
+    /** The model's confidence that the text has been classified correctly */
+    classificationConfidence: number | null;
+    /** If the document was submitted as an image, this is the confidence that the text in the image has been correctly read by the model. */
+    textExtractionConfidence: number | null;
     isVerified: boolean;
     isClientVerified?: boolean;
     isAutoVerified?: boolean;
@@ -2004,6 +2009,15 @@ export declare interface ResumeSearchConfig {
      * NOTE: This property will not be serialized. It can only be populated by the server.
      */
     readonly username?: string;
+    /** A list of actions to show in the dropdown in the embedded search tool */
+    actions?: ResumeSearchConfigActionsItem[];
+}
+
+export declare interface ResumeSearchConfigActionsItem {
+    /** Human readable label to display in the UI */
+    label?: string;
+    /** Name of the event to be triggered */
+    eventName?: string;
 }
 
 export declare interface ResumeSearchDetail {