+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add actions to embedded resume search
+- Add actions to embedded search
+- Add actions to resume search
 
 ## [2.5.1] - 2022-12-09
 ### Changed